requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio==3.4.3
asynctest==0.13.0
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
Deprecated==1.2.14
Flask==2.0.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
flask-io==1.14.3
Flask-Limiter==2.2.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
MarkupSafe==2.1.3
marshmallow==2.21.0
multidict==6.0.4
packaging==23.2
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.27.1
six==1.16.0
twitchAPI==4.0.1
typing-extensions==4.7.1
urllib3==1.26.18
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
